+++ web/templates/edited_notice.hbs
@@ -1,14 +1,14 @@
 {{#if modified}}
     {{#if msg/local_edit_timestamp}}
-        <div class="message_edit_notice" data-tippy-content="{{t 'Last edited {last_edit_timestr}.'}}">
+        <div class="message_edit_notice">
             {{t "SAVING"}}
         </div>
     {{else if moved}}
-        <div class="message_edit_notice" data-tippy-content="{{t 'Last moved {last_edit_timestr}.'}}">
+        <div class="message_edit_notice">
             {{t "MOVED"}}
         </div>
     {{else}}
-        <div class="message_edit_notice" data-tippy-content="{{t 'Last edited {last_edit_timestr}.'}}">
+        <div class="message_edit_notice">
             {{t "EDITED"}}
         </div>
     {{/if}}
